# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: VPS Billing Platform API
  version: 2.0.0
  license: {name: Proprietary}
servers:
  - url: /api/v1
security: [{userSession: []}]
paths:
...
  /orders:
    get:
      summary: List user orders
      operationId: listOrders
      responses:
        "200": {description: Orders}
        "401": {$ref: "#/components/responses/Unauthenticated"}
    post:
      summary: Create an order
      operationId: createOrder
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plan_id, quantity]
              properties:
                plan_id: {type: string, format: uuid}
                quantity: {type: integer, minimum: 1}
                promotion_code: {type: string, maxLength: 64}
      responses:
        "201": {description: "Order, invoice, and pending payment created"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "422": {$ref: "#/components/responses/ValidationFailure"}
  /invoices:
    get:
      summary: List user invoices
      operationId: listInvoices
      responses:
        "200": {description: Invoices}
        "401": {$ref: "#/components/responses/Unauthenticated"}
  /wallet:
    get:
      summary: Get or initialize a zero-balance wallet
      operationId: getWallet
      parameters:
        - name: currency
          in: query
          schema: {type: string, minLength: 3, maxLength: 3, default: USD}
      responses:
        "200": {description: Wallet}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "422": {$ref: "#/components/responses/ValidationFailure"}
  /webhooks/payments/fake:
    post:
      summary: Receive a signed fake payment callback
      operationId: fakePaymentWebhook
      security: []
      parameters:
        - name: X-Fake-Signature
          in: header
          required: true
          schema: {type: string}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/PaymentWebhook"}
      responses:
        "200": {description: Processed or already processed idempotently}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "404": {description: Payment not found}
        "409": {description: Payment state conflict}
        "422": {$ref: "#/components/responses/ValidationFailure"}
...
  /billing-profile:
    get: {summary: Get the authenticated user's tax and billing profile, responses: {"200": {description: Billing profile}}}
    put: {summary: Update the authenticated user's tax and billing profile, parameters: [{$ref: "#/components/parameters/CSRFToken"}], responses: {"200": {description: Updated billing profile}, "422": {$ref: "#/components/responses/ValidationFailure"}}}
...
components:
  securitySchemes:
    userSession: {type: apiKey, in: cookie, name: vps_user_session}
    adminSession: {type: apiKey, in: cookie, name: vps_admin_session}
...
  parameters:
...
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: {type: string}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 8, maxLength: 255}
...
  responses:
    Unauthenticated: {description: Authentication required, content: {application/json: {schema: {$ref: "#/components/schemas/ErrorResponse"}}}}
    Forbidden: {description: Permission or CSRF validation failed, content: {application/json: {schema: {$ref: "#/components/schemas/ErrorResponse"}}}}
...
    ValidationFailure: {description: Validation failure, content: {application/json: {schema: {$ref: "#/components/schemas/ErrorResponse"}}}}
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [success, error, request_id]
      properties:
        success: {type: boolean, const: false}
        request_id: {type: string}
        error:
          type: object
          required: [code, message_key]
          properties:
            code: {type: string}
            message_key: {type: string}
            details: {type: object, additionalProperties: true}
    PaymentWebhook:
      type: object
      additionalProperties: false
      required: [event_id, payment_id, external_payment_id, status, amount_minor, currency]
      properties:
        event_id: {type: string, minLength: 1, maxLength: 255}
        payment_id: {type: string, format: uuid}
        external_payment_id: {type: string, minLength: 1, maxLength: 255}
        status: {type: string, enum: [succeeded]}
        amount_minor: {type: integer, format: int64, minimum: 0}
        currency: {type: string, minLength: 3, maxLength: 3}
...